1. Implemented Python's standard ConfigParser library to manage configuration instead of the custom method being used currently 2. Replaced the mergeArgs() with getConfig() method to 	a. First check if mandatory arguments, lat, lon, method are stored in settings.ini 	b. If not, look at the args provided during startup 	c. Set the args and return it 3. Changed job commands to make it fully compatible with python3 4. Improved the formatting of output to be more descriptive 5. Added a method to play surah baqarah on fridays as 8 am 6. Updated README
