+++ .gitignore
@@ -2,4 +2,5 @@
 *.log
 *.pyc
 adhan.log
-.settings
+settings.ini
+__pycache__
